// src/core/models/enums/app-color.ts
export enum AppColor {
    blue = '#157DF1',
    darkBlue = '#0057B9',
    almostWhiteBlue = '#ABD2FF',
    white = '#FFFFFF',
    black = '#333333',
    grey = '#808080',
    red = '#F0425F',
    contrastGrey = '#E8E9EB',
    blueAlternative = '#509DF4',
    whiteBlue = '#C5DFFB',
    lightBlue = '#8ABEF8',
}